+++ src/app/website/documentation/rest-api/rest-rdfsource/RESTRdfSourceView.ts
@@ -1,8 +1,8 @@
-import {Component, ElementRef} from "@angular/core";
-import {CORE_DIRECTIVES} from "@angular/common";
-import {Title} from "@angular/platform-browser";
+import { Component, ElementRef } from "@angular/core";
+import { CORE_DIRECTIVES } from "@angular/common";
+import { Title } from "@angular/platform-browser";
 
-import HighlightDirective from "app/directives/HighlightDirective";
+import { HighlightDirective } from "carbon-panel/directives/highlight.directive";
 
 import SidebarComponent from "./../../sidebar/SidebarComponent";